validator/pyproject.toml
[project]
name = "djinn-validator"
version = "0.1.0"
description = "Djinn Protocol Bittensor validator — manages key shares, MPC, outcome attestation, miner scoring"
requires-python = ">=3.11"
dependencies = [
    "bittensor>=8.0.0,<11",
    "fastapi>=0.110.0,<1",
    "uvicorn>=0.30.0,<1",
    "pydantic>=2.7.0,<3",
    "web3>=7.0.0,<8",
    "httpx>=0.27.0,<1",
    "python-dotenv>=1.0.0,<2",
    "structlog>=24.1,<26",
    "prometheus-client>=0.21.0,<1",
    "PyJWT[crypto]>=2.8.0,<3",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0,<9",
    "pytest-asyncio>=0.24.0,<1",
    "pytest-httpx~=0.35.0",
    "coverage~=7.0.0",
    "ruff~=0.8.0",
]

[project.scripts]
djinn-validator = "djinn_validator.main:main"

[tool.ruff]
target-version = "py311"
line-length = 120

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]
ignore = ["E501", "E402", "N806"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
